Maya plug-in pre-alpha (continued)
All 68 data points are now being taken into account within the plug-in;
I've added a few new features too:

	- Anchor locally: calculates translation coordinates based on the
		deltas from a "neutral" position, defined below.
	- Neutral face location: allows user to define a "neutral" face
		size/location. Can be re-calibrated at any time with the
		calibrateNeutralFace() MEL call.
	- Amplify amount: specify what the incoming coordinate data should
		be multiplied by, for a higher "sensitivity".

Within the Stringless server, I've changed FaceDetector's outline
drawing functionality to display circles on each point instead of the
original lines, so that the connection between the points in the server
and in Maya are much clearer.

Things to think about doing:
	
	- Fix the calibrateNeutralFace() function so it works every time
		(my guess is it's asynchronous, which is causing the issue?)
	- Add feature to update neutralFace's origin as user moves face
		around (follow user's face, probably best to do via the nose)
	- Re-write the MEL script into Python, for scalability/double
		support/editor support/etc. (at some point in the future)
	- Author compile instructions for alpha release
	- Get some sleep

+++ src/FaceDetector.h
@@ -21,19 +21,19 @@ namespace Stringless {
 
 class FaceDetector {
 public:
-    FaceDetector(int camera_number, 
-                 int downsample_ratio, 
-                 int sample_rate, 
-                 char *face_landmarks_location);
+    FaceDetector(int    camera_number, 
+                 double downsample_ratio, 
+                 int    sample_rate, 
+                 char   *face_landmarks_location);
     FaceDetector(const FaceDetector& orig);
     virtual ~FaceDetector();
     
     int start(Writer &writer);
 private:
-    int     camera_number;
-    int     downsample_ratio;
-    int     sample_rate;
-    char    *face_landmarks_location;
+    int         camera_number;
+    double      downsample_ratio;
+    int         sample_rate;
+    char        *face_landmarks_location;
 
 };
